feat: Add enrollApi.ts for enrolling and unenrolling students
This commit adds the `enrollApi.ts` file in the `api/interactions` folder. It contains functions for enrolling and unenrolling students from subjects. The functions `enrollStudent`, `unenrollStudent`, and `getEnrolledSubjects` are implemented using Axios to make API requests to the server. These functions will be used to handle student enrollment and unenrollment in the application.