feat: Refactor StatusActions component to improve node disabling logic
The code changes in the `StatusActions.tsx` file refactor the `StatusActions` component in the `graph/behaviors` folder. The changes improve the logic for disabling nodes and add the ability to disable nodes based on certain conditions. The `disableViewedNode` function now accepts an additional parameter `disabledNodes` to keep track of disabled nodes. This enhances the functionality of the component and ensures that nodes are properly disabled when necessary.