refactor: Update GoogleLogin component to use Skeleton for loading state
The GoogleLogin component has been updated to use the Skeleton component from the @ui/skeleton module for the loading state. This improves the user experience by providing a visual placeholder while the authentication is being processed. The Skeleton component displays a pulsating animation to indicate that content is loading. This change replaces the previous usage of the Spinner component.