refactor: Improve touch event handling in MenuActions component
The touch event handling in the MenuActions component has been improved to ensure that the menu closes properly. Previously, the menu would not close when dragging on the canvas, leading to a poor user experience. This fix addresses the issue by adding a new touchend event listener on the window object and modifying the handleNodeTouchEnd function to accept a parameter indicating whether the menu should close. Aso, now the mantain the touch to open the menu on mobile works