feat: Add behavior to update node status on graph change
This commit adds a new behavior component, UpdateNodeStatusOnGraphChange, which is responsible for updating the status of nodes in the graph when there is a change. It listens for changes in the graph data and updates the state of each node accordingly. This behavior improves the overall functionality and user experience of the graph component.