refactor: Add createCareer2 route for creating a new career
This commit adds a new route, createCareer2, to the careers.go file. The route allows for the creation of a new career using the provided career form data. This change enhances the functionality of the careers module and improves the user experience by providing a way to create careers with specific requirements.