feat: Add createCareers function for creating new careers
This commit adds a new function, createCareers, to the careersApi.ts file. The function allows for the creation of new careers by making a POST request to the "/careers/prueba/" endpoint with the provided data. This feature enhances the functionality of the careers module and improves the user experience by providing a way to create careers with specific requirements.