feat: Add getSubjectsGraph function for fetching subject graph
This commit adds a new function, getSubjectsGraph, to the subjectsAPI.ts file. The function fetches the subject graph data by making a GET request to the "/subjects/graph/" endpoint with the provided career parameter. This feature enhances the functionality of the code by allowing the retrieval of subject graph data for a specific career.