refactor: Update FloatingLabelInput component to use FormField and FormControl
This commit refactors the FloatingLabelInput component in the floating-label-input.tsx file. The changes include replacing the usage of the deprecated Input and Label components with the new FormField and FormControl components from the @ui/form module. This refactor enhances the maintainability and compatibility of the code by using the latest form-related components and APIs.